tests/justhtml-tests/in_body_recovery.dat

#data
<dl><dd>x</dd></dl>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "x"

#data
<p><u><section><meta>
#errors
(1,3): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <p>
|       <u>
|     <section>
|       <meta>

#data
<marquee></body><!
#errors
(1,9): expected-doctype-but-got-start-tag
(1,10): unexpected-end-tag
(1,18): eof-in-tag

#document
| <html>
|   <head>
|   <body>
|     <marquee>
|       <!--  -->
